requirements.txt
torchvision==0.11.1
seaborn==0.8.1
numpy==1.16.2
scipy==1.3.1
setproctitle==1.1.10
matplotlib==3.0.3
pandas==0.24.2
statsmodels==0.8.0
opencv_python==3.4.0.12
torch==1.10.0
Pillow==7.0.0
python_dateutil==2.8.1
setGPU==0.0.7
pytorch-geometric
